Changed to f32 usage to allow for even better efficiency :()

+++ src/network/activations.rs
@@ -1,10 +1,10 @@
-use std::f64::consts::E;
+use std::f32::consts::E;
 use serde::{Deserialize, Serialize};
 
 #[derive(Clone)]
 pub struct Activation<'a>{
-    pub function: &'a dyn Fn(f64) -> f64,
-    pub derivative: &'a dyn Fn(f64) -> f64
+    pub function: &'a dyn Fn(f32) -> f32,
+    pub derivative: &'a dyn Fn(f32) -> f32
 }
 
 #[derive(Clone, Serialize, Deserialize)]
@@ -34,10 +34,10 @@ const SIGMOID: Activation = Activation {
 
 const TANH: Activation = Activation {
     function: &|x| {
-        let res = f64::tanh(x);
+        let res = f32::tanh(x);
         return res;
     },
-    derivative: &|x| 1.0 - f64::tanh(x).powf(2.0)
+    derivative: &|x| 1.0 - f32::tanh(x).powf(2.0)
 };
 
 const RELU: Activation = Activation {
